% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouseMaze.R
\name{mouse}
\alias{mouse}
\title{Mouse in the maze}
\usage{
mouse(n)
}
\arguments{
\item{n}{The length of maze.}
}
\value{
You can know what the path does mouse go throuth and what the path length is.
}
\description{
This function can help the mouse finding its way to escape the maze from 1 to n^2.
We ristrict the maze only can be n^2 size maze, and n>1.
}
\examples{
mouse(3)
}
